Further seperate out methods
