+++ x/gamm/keeper/swap.go
@@ -1,7 +1,10 @@
 package keeper
 
 import (
+	"errors"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/osmosis-labs/osmosis/v12/x/gamm/keeper/internal/events"
 	"github.com/osmosis-labs/osmosis/v12/x/gamm/types"
@@ -42,18 +45,59 @@ func (k Keeper) swapExactAmountIn(
 	tokenOutMinAmount sdk.Int,
 	swapFee sdk.Dec,
 ) (tokenOutAmount sdk.Int, err error) {
-	pool, tokenOut, err := k.estimateSwapExactAmountIn(ctx, pool, tokenIn, tokenOutDenom, tokenOutMinAmount, swapFee)
+	pool, tokenOut, err := k.swapExactAmountInNoTokenSend(ctx, pool, tokenIn, tokenOutDenom, tokenOutMinAmount, swapFee)
 	if err != nil {
 		return sdk.Int{}, err
 	}
 
 	// Settles balances between the tx sender and the pool to match the swap that was executed earlier.
 	// Also emits swap event and updates related liquidity metrics
-	if err := k.updatePoolForSwap(ctx, pool, sender, tokenIn, tokenOut); err != nil {
+	if err := k.updatePoolForSwap(ctx, pool, tokenIn, tokenOut); err != nil {
+		return sdk.Int{}, err
+	}
+
+	err = k.sendCoinsAfterSwap(ctx, pool, tokenIn, tokenOut, sender)
+	if err != nil {
 		return sdk.Int{}, err
 	}
 
 	return tokenOut.Amount, nil
+}
+
+// swapExactAmountInNoTokenSend performs `SwapOutAmtGivenIn` for the given inputs,
+// and potentially mutates state. It does not save the new pool struct, or do token transfers.
+// Returns the altered pool after swap, and the token out when swapped.
+func (k Keeper) swapExactAmountInNoTokenSend(
+	ctx sdk.Context,
+	pool types.PoolI,
+	tokenIn sdk.Coin,
+	tokenOutDenom string,
+	tokenOutMinAmount sdk.Int,
+	swapFee sdk.Dec,
+) (updatedPool types.PoolI, tokenOut sdk.Coin, err error) {
+	if tokenIn.Denom == tokenOutDenom {
+		return pool, sdk.Coin{}, errors.New("cannot trade same denomination in and out")
+	}
+	tokensIn := sdk.Coins{tokenIn}
+
+	// Executes the swap in the pool and stores the output. Updates pool assets but
+	// does not actually transfer any tokens to or from the pool.
+	tokenOutCoin, err := pool.SwapOutAmtGivenIn(ctx, tokensIn, tokenOutDenom, swapFee)
+	if err != nil {
+		return pool, sdk.Coin{}, err
+	}
+
+	tokenOutAmount := tokenOutCoin.Amount
+
+	if !tokenOutAmount.IsPositive() {
+		return pool, sdk.Coin{}, sdkerrors.Wrapf(types.ErrInvalidMathApprox, "token amount must be positive")
+	}
+
+	if tokenOutAmount.LT(tokenOutMinAmount) {
+		return pool, sdk.Coin{}, sdkerrors.Wrapf(types.ErrLimitMinAmount, "%s token is lesser than min amount", tokenOutDenom)
+	}
+
+	return pool, tokenOutCoin, nil
 }
 
 func (k Keeper) SwapExactAmountOut(
@@ -85,12 +129,17 @@ func (k Keeper) swapExactAmountOut(
 	tokenOut sdk.Coin,
 	swapFee sdk.Dec,
 ) (tokenInAmount sdk.Int, err error) {
-	pool, tokenIn, err := k.estimateSwapExactAmountOut(ctx, pool, tokenInDenom, tokenInMaxAmount, tokenOut, swapFee)
+	pool, tokenIn, err := k.swapExactAmountOutNoTokenSend(ctx, pool, tokenInDenom, tokenInMaxAmount, tokenOut, swapFee)
 	if err != nil {
 		return sdk.Int{}, err
 	}
 
-	err = k.updatePoolForSwap(ctx, pool, sender, tokenIn, tokenOut)
+	err = k.updatePoolForSwap(ctx, pool, tokenIn, tokenOut)
+	if err != nil {
+		return sdk.Int{}, err
+	}
+
+	err = k.sendCoinsAfterSwap(ctx, pool, tokenIn, tokenOut, sender)
 	if err != nil {
 		return sdk.Int{}, err
 	}
@@ -98,13 +147,48 @@ func (k Keeper) swapExactAmountOut(
 	return tokenIn.Amount, nil
 }
 
-// updatePoolForSwap takes a pool, sender, and tokenIn, tokenOut amounts
-// It then updates the pool's balances to the new reserve amounts, and
-// sends the in tokens from the sender to the pool, and the out tokens from the pool to the sender.
+// swapExactAmountOutNoTokenSend performs `SwapInAmtGivenOut` for the given inputs,
+// and potentially mutates state. It does not save the new pool struct, or do token transfers.
+// Returns the altered pool after swap, and the token in when swapped.
+func (k Keeper) swapExactAmountOutNoTokenSend(
+	ctx sdk.Context,
+	pool types.PoolI,
+	tokenInDenom string,
+	tokenInMaxAmount sdk.Int,
+	tokenOut sdk.Coin,
+	swapFee sdk.Dec,
+) (updatedPool types.PoolI, tokenIn sdk.Coin, err error) {
+	if tokenInDenom == tokenOut.Denom {
+		return pool, sdk.Coin{}, errors.New("cannot trade same denomination in and out")
+	}
+
+	poolOutBal := pool.GetTotalPoolLiquidity(ctx).AmountOf(tokenOut.Denom)
+	if tokenOut.Amount.GTE(poolOutBal) {
+		return pool, sdk.Coin{}, sdkerrors.Wrapf(types.ErrTooManyTokensOut,
+			"can't get more tokens out than there are tokens in the pool")
+	}
+
+	tokenIn, err = pool.SwapInAmtGivenOut(ctx, sdk.Coins{tokenOut}, tokenInDenom, swapFee)
+	if err != nil {
+		return pool, sdk.Coin{}, err
+	}
+	tokenInAmount := tokenIn.Amount
+
+	if tokenInAmount.LTE(sdk.ZeroInt()) {
+		return pool, sdk.Coin{}, sdkerrors.Wrapf(types.ErrInvalidMathApprox, "token amount is zero or negative")
+	}
+
+	if tokenInAmount.GT(tokenInMaxAmount) {
+		return pool, sdk.Coin{}, sdkerrors.Wrapf(types.ErrLimitMaxAmount, "Swap requires %s, which is greater than the amount %s", tokenIn, tokenInMaxAmount)
+	}
+
+	return pool, tokenIn, nil
+}
+
+// updatePoolForSwap updates the pool balance and the pool state.
 func (k Keeper) updatePoolForSwap(
 	ctx sdk.Context,
 	pool types.PoolI,
-	sender sdk.AccAddress,
 	tokenIn sdk.Coin,
 	tokenOut sdk.Coin,
 ) error {
@@ -116,7 +200,22 @@ func (k Keeper) updatePoolForSwap(
 		return err
 	}
 
-	err = k.bankKeeper.SendCoins(ctx, sender, pool.GetAddress(), sdk.Coins{
+	k.RecordTotalLiquidityIncrease(ctx, tokensIn)
+	k.RecordTotalLiquidityDecrease(ctx, tokensOut)
+
+	return nil
+}
+
+// sendCoinsAfterSwap sends the coins after a swap has been executed.
+// this method also includes hooks and emitting events.
+func (k Keeper) sendCoinsAfterSwap(
+	ctx sdk.Context,
+	pool types.PoolI,
+	tokenIn sdk.Coin,
+	tokenOut sdk.Coin,
+	sender sdk.AccAddress,
+) error {
+	err := k.bankKeeper.SendCoins(ctx, sender, pool.GetAddress(), sdk.Coins{
 		tokenIn,
 	})
 	if err != nil {
@@ -130,10 +229,10 @@ func (k Keeper) updatePoolForSwap(
 		return err
 	}
 
+	tokensIn := sdk.Coins{tokenIn}
+	tokensOut := sdk.Coins{tokenOut}
 	events.EmitSwapEvent(ctx, sender, pool.GetId(), tokensIn, tokensOut)
 	k.hooks.AfterSwap(ctx, sender, pool.GetId(), tokensIn, tokensOut)
-	k.RecordTotalLiquidityIncrease(ctx, tokensIn)
-	k.RecordTotalLiquidityDecrease(ctx, tokensOut)
 
-	return err
+	return nil
 }
